delete fresherjob code reduced
