+++ src/utils/UI/FresherJobCard/FresherJobCard.js
@@ -6,8 +6,40 @@ import { UserContext } from "../../../App";
 import axios from "axios";
 import toast, { Toaster } from "react-hot-toast";
 
-const FresherJobCard = ({ fresherjob, deletePost, userId }) => {
+const FresherJobCard = ({ fresherjob, userId }) => {
   const { state, dispatch } = useContext(UserContext);
+
+  const deletePost = (postId) => {
+    axios({
+      method: "delete",
+      url: "http://localhost:5000/employer/delete-freshersjob",
+      data: {
+        postId,
+      },
+      headers: {
+        Authorization: "Bearer " + localStorage.getItem("jwt"),
+        "Content-Type": "application/json",
+      },
+    })
+      .then((res) => {
+        console.log(res);
+        if (res.data.error) {
+          console.log(res.data.error);
+          // alert(res.data.error);
+          const notify = () => toast(res.data.error);
+          notify();
+        } else {
+          // console.log(res.data.jobs);
+          // setJobs(res.data.jobs);
+          // console.log(jobs);
+          const notify = () => toast(res.data.message);
+          notify();
+        }
+      })
+      .catch((err) => {
+        console.log("Error: ", err);
+      });
+  };
 
   const bookMarkPost = (postId) => {
     axios({
